requirements.txt
requests
beautifulsoup4
twilio
openai
apscheduler
python-dotenv
